voice-assistant/requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
uvicorn==0.32.1
fastapi==0.115.6
websockets==14.1
pydantic==2.10.3
python-multipart==0.0.20
